Merges into request
