+++ console/program/src/lib.rs
@@ -23,9 +23,6 @@ extern crate enum_index_derive;
 pub use snarkvm_console_network::Network;
 pub use snarkvm_console_types::prelude::*;
 
-mod call;
-pub use call::*;
-
 mod data;
 pub use data::*;
