--- console/program/src/request/to_fields.rs
@@ -1,40 +1,0 @@
-// Copyright (C) 2019-2022 Aleo Systems Inc.
-// This file is part of the snarkVM library.
-
-// The snarkVM library is free software: you can redistribute it and/or modify
-// it under the terms of the GNU General Public License as published by
-// the Free Software Foundation, either version 3 of the License, or
-// (at your option) any later version.
-
-// The snarkVM library is distributed in the hope that it will be useful,
-// but WITHOUT ANY WARRANTY; without even the implied warranty of
-// MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE. See the
-// GNU General Public License for more details.
-
-// You should have received a copy of the GNU General Public License
-// along with the snarkVM library. If not, see <https://www.gnu.org/licenses/>.
-
-use super::*;
-
-impl<N: Network> ToFields for Request<N> {
-    type Field = Field<N>;
-
-    /// Returns this request as a list of field elements.
-    fn to_fields(&self) -> Result<Vec<Self::Field>> {
-        // Initialize a vector for the list of field elements.
-        let mut fields = Vec::with_capacity(4 + self.inputs.len());
-        fields.push(self.caller.to_field()?);
-        fields.push(self.network_id.to_field()?);
-        fields.extend(self.program_id.to_fields()?);
-        fields.push(self.function_name.to_field()?);
-        fields.extend(
-            self.inputs.iter().map(|input| input.to_fields()).collect::<Result<Vec<_>>>()?.into_iter().flatten(),
-        );
-
-        // Ensure the number of field elements does not exceed the maximum allowed size.
-        match fields.len() <= N::MAX_DATA_SIZE_IN_FIELDS as usize {
-            true => Ok(fields),
-            false => bail!("Request exceeds maximum allowed size"),
-        }
-    }
-}
